pyproject.toml
[project]
name = "g-league-scouting"
version = "0.1.0"
description = "G-League scouting platform using NBA and G-League data"
authors = [
    {name = "Julien Jaffré"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "nba-api (>=1.10.0,<2.0.0)",
    "google-cloud-storage (>=2.10,<3.0)",
    "dbt-bigquery (>=1.10.0,<2.0.0)",
    "google-cloud-bigquery (>=3.35.1,<4.0.0)",
    "streamlit (>=1.47.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
matplotlib = "^3.10.3"

